refactor(suite-native): move suite sync hooks to suite-sync
