+++ suite-native/labeling/package.json
@@ -13,26 +13,19 @@
     },
     "dependencies": {
         "@gorhom/bottom-sheet": "5.2.8",
-        "@react-navigation/native": "7.1.28",
         "@reduxjs/toolkit": "2.11.2",
         "@suite-common/device": "workspace:*",
         "@suite-common/message-system": "workspace:*",
         "@suite-common/suite-sync": "workspace:*",
-        "@suite-common/suite-sync-storage": "workspace:*",
-        "@suite-common/suite-sync-types": "workspace:*",
         "@suite-common/validators": "workspace:*",
         "@suite-common/wallet-core": "workspace:*",
         "@suite-common/wallet-utils": "workspace:*",
-        "@suite-native/alerts": "workspace:*",
         "@suite-native/atoms": "workspace:*",
         "@suite-native/forms": "workspace:*",
         "@suite-native/intl": "workspace:*",
-        "@suite-native/navigation": "workspace:*",
-        "@suite-native/services": "workspace:*",
         "@suite-native/settings": "workspace:*",
-        "@suite-native/toasts": "workspace:*",
+        "@suite-native/suite-sync": "workspace:*",
         "@trezor/connect": "workspace:*",
-        "@trezor/type-utils": "workspace:*",
         "react": "19.2.4",
         "react-native": "0.83.2",
         "react-redux": "9.2.0"
