+++ suite-native/module-accounts-management/package.json
@@ -55,6 +55,7 @@
         "@suite-native/navigation": "workspace:*",
         "@suite-native/qr-code": "workspace:*",
         "@suite-native/services": "workspace:*",
+        "@suite-native/suite-sync": "workspace:*",
         "@suite-native/toasts": "workspace:*",
         "@suite-native/tokens": "workspace:*",
         "@suite-native/transactions": "workspace:*",
