+++ suite-native/module-accounts-management/src/components/AccountRenameForm.tsx
@@ -18,12 +18,9 @@ import { Box, Button, type InputType, VStack } from '@suite-native/atoms';
 import { featureUsed } from '@suite-native/feature-feedback';
 import { Form, TextInputField } from '@suite-native/forms';
 import { Translation, useTranslate } from '@suite-native/intl';
-import {
-    type CombinedLabelingState,
-    selectIsLabellingAllowed,
-    useSuiteSyncErrorHandler,
-} from '@suite-native/labeling';
+import { type CombinedLabelingState, selectIsLabellingAllowed } from '@suite-native/labeling';
 import { useNativeServices } from '@suite-native/services';
+import { useSuiteSyncErrorHandler } from '@suite-native/suite-sync';
 
 type AccountRenameFormProps = {
     accountKey: AccountKey;
