+++ suite-native/module-home/package.json
@@ -41,11 +41,11 @@
         "@suite-native/firmware": "workspace:*",
         "@suite-native/graph": "workspace:*",
         "@suite-native/intl": "workspace:*",
-        "@suite-native/labeling": "workspace:*",
         "@suite-native/link": "workspace:*",
         "@suite-native/navigation": "workspace:*",
         "@suite-native/services": "workspace:*",
         "@suite-native/settings": "workspace:*",
+        "@suite-native/suite-sync": "workspace:*",
         "@suite-native/toasts": "workspace:*",
         "@trezor/device-utils": "workspace:*",
         "@trezor/styles-native": "workspace:*",
