+++ suite-native/module-settings/package.json
@@ -50,6 +50,7 @@
         "@suite-native/services": "workspace:*",
         "@suite-native/settings": "workspace:*",
         "@suite-native/storage": "workspace:*",
+        "@suite-native/suite-sync": "workspace:*",
         "@suite-native/theme": "workspace:*",
         "@suite-native/toasts": "workspace:*",
         "@suite-native/trading-residence": "workspace:*",
