+++ suite-native/suite-sync/src/index.ts
@@ -1,1 +1,3 @@
 export { createSuiteSyncNativeCompositionRoot } from './createSuiteSyncNativeCompositionRoot';
+export { useSuiteSyncErrorHandler } from './useSuiteSyncErrorHandler';
+export { useTurnOnSuiteSyncGuard } from './useTurnOnSuiteSyncGuard';
