+++ suite-native/suite-sync/tsconfig.json
@@ -1,9 +1,17 @@
 {
     "extends": "../../tsconfig.base.json",
-    "compilerOptions": { "outDir": "libDev" },
+    "compilerOptions": {
+        "outDir": "libDev"
+    },
     "references": [
         {
+            "path": "../../suite-common/device"
+        },
+        {
             "path": "../../suite-common/delegated-identity-key-types"
+        },
+        {
+            "path": "../../suite-common/message-system"
         },
         {
             "path": "../../suite-common/platform-encryption"
@@ -16,10 +24,31 @@
         },
         {
             "path": "../../suite-common/suite-sync-quota-manager"
+            "path": "../../suite-common/suite-sync-storage"
         },
         {
             "path": "../../suite-common/suite-sync-types"
         },
-        { "path": "../../packages/connect" }
+        {
+            "path": "../alerts"
+        },
+        {
+            "path": "../intl"
+        },
+        {
+            "path": "../navigation"
+        },
+        {
+            "path": "../services"
+        },
+        {
+            "path": "../toasts"
+        },
+        {
+            "path": "../../packages/connect"
+        },
+        {
+            "path": "../../packages/type-utils"
+        }
     ]
 }
