+++ suite-native/transactions/package.json
@@ -38,6 +38,7 @@
         "@suite-native/navigation": "workspace:*",
         "@suite-native/scrollview": "workspace:*",
         "@suite-native/services": "workspace:*",
+        "@suite-native/suite-sync": "workspace:*",
         "@suite-native/tokens": "workspace:*",
         "@trezor/blockchain-link-types": "workspace:*",
         "@trezor/connect": "workspace:*",
